refactor(channel): extract lookups + composer input helpers
Phase 2 continuation. Two cohesive extractions from channel.go:

(1) channel_lookups.go (90 LOC, new): findTaskByID, findRequestByID,
    focusRequest, answerRequest, openTaskActionPicker,
    openRequestActionPicker. Pure ID-to-entity finds plus the
    request-focus / answer-flow transitions and picker bridges.

(2) channel_input.go (143 LOC, new): updateOverlaysForInput +
    updateInputOverlays + updateThreadOverlays +
    updateOverlaysForCurrentInput, setActiveInput, activeInputString,
    insertAcceptedMention, restoreMainSnapshot, restoreThreadSnapshot,
    composerMotionKey, composerInsertRunes. Composer overlay management
    + the cursor-motion/literal-insert key classification.

  - cmd/wuphf/channel.go: -192 LOC (3049 -> 2857)
  - cmd/wuphf/channel_lookups.go: +90 LOC (new)
  - cmd/wuphf/channel_input.go: +143 LOC (new)

Cumulative channel.go reduction across Phase 2:
  4807 -> 4687 (pickers slice 1)        -120
       -> 4463 (sidebar)                -224
       -> 4249 (mouse + geometry)       -214
       -> 3551 (commands)               -698
       -> 3049 (pickers + client)       -502
       -> 2857 (lookups + composer)     -192
  Net: -1950 LOC; channel.go shrunk by 41%.

No behavior change. cmd/wuphf tests pass.
