+++ cmd/wuphf/channel.go
@@ -2729,200 +2729,10 @@ func (m channelModel) updateThread(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
 
 // Picker option builders moved to channel_pickers.go.
 
-func (m channelModel) findTaskByID(id string) (channelui.Task, bool) {
-	for _, task := range m.tasks {
-		if task.ID == id {
-			return task, true
-		}
-	}
-	return channelui.Task{}, false
-}
+// Lookup helpers + request-action dispatch moved to channel_lookups.go.
 
-func (m channelModel) findRequestByID(id string) (channelui.Interview, bool) {
-	for _, req := range m.requests {
-		if req.ID == id {
-			return req, true
-		}
-	}
-	return channelui.Interview{}, false
-}
-
-func (m channelModel) focusRequest(req channelui.Interview, notice string) (tea.Model, tea.Cmd) {
-	if req.Blocking || req.Required {
-		m.activeApp = channelui.OfficeAppMessages
-	} else {
-		m.activeApp = channelui.OfficeAppRequests
-	}
-	m.syncSidebarCursorToActive()
-	m.pending = &req
-	m.selectedOption = m.recommendedOptionIndex()
-	m.notice = notice
-	if req.ReplyTo != "" {
-		m.threadPanelOpen = true
-		m.threadPanelID = req.ReplyTo
-	}
-	return m, tea.Batch(pollRequests(m.activeChannel))
-}
-
-func (m channelModel) answerRequest(req channelui.Interview) (tea.Model, tea.Cmd) {
-	if req.Blocking || req.Required {
-		m.activeApp = channelui.OfficeAppMessages
-	} else {
-		m.activeApp = channelui.OfficeAppRequests
-	}
-	m.syncSidebarCursorToActive()
-	m.pending = &req
-	m.selectedOption = m.recommendedOptionIndex()
-	m.notice = "Answering request " + req.ID + ". Type your answer and press Enter."
-	if req.ReplyTo != "" {
-		m.threadPanelOpen = true
-		m.threadPanelID = req.ReplyTo
-	}
-	return m, nil
-}
-
-func (m *channelModel) openTaskActionPicker(task channelui.Task) tea.Cmd {
-	actions := m.buildTaskActionPickerOptions(task)
-	if len(actions) == 0 {
-		return nil
-	}
-	m.picker = tui.NewPicker("Task: "+channelui.TruncateText(task.Title, 40), actions)
-	m.picker.SetActive(true)
-	m.pickerMode = channelPickerTaskAction
-	m.notice = "Choose a task action."
-	return nil
-}
-
-func (m *channelModel) openRequestActionPicker(req channelui.Interview) tea.Cmd {
-	actions := m.buildRequestActionPickerOptions(req)
-	if len(actions) == 0 {
-		return nil
-	}
-	m.picker = tui.NewPicker("Request: "+channelui.TruncateText(req.TitleOrQuestion(), 40), actions)
-	m.picker.SetActive(true)
-	m.pickerMode = channelPickerRequestAction
-	m.notice = "Choose a request action."
-	return nil
-}
-
-func (m *channelModel) updateOverlaysForInput(input []rune, cursor int) {
-	text := string(input)
-	if cursor < 0 {
-		cursor = 0
-	}
-	if cursor > len(input) {
-		cursor = len(input)
-	}
-	m.autocomplete.UpdateQuery(strings.TrimLeft(text[:cursor], " "))
-	m.mention.UpdateAgents(channelMentionAgents(m.members))
-	m.mention.UpdateQuery(text[:cursor])
-}
-
-func (m *channelModel) updateInputOverlays() {
-	m.updateOverlaysForInput(m.input, m.inputPos)
-}
-
-func (m *channelModel) updateThreadOverlays() {
-	m.updateOverlaysForInput(m.threadInput, m.threadInputPos)
-}
-
-func (m *channelModel) updateOverlaysForCurrentInput() {
-	if m.focus == focusThread && m.threadPanelOpen {
-		m.updateThreadOverlays()
-		return
-	}
-	if m.focus == focusMain {
-		m.updateInputOverlays()
-		m.maybeActivateChannelPickerFromInput()
-		return
-	}
-	m.autocomplete.Dismiss()
-	m.mention.Dismiss()
-}
-
-func (m *channelModel) setActiveInput(text string) {
-	if m.focus == focusThread && m.threadPanelOpen {
-		m.threadInput = []rune(text)
-		m.threadInputPos = len(m.threadInput)
-		m.threadInputHistory.ResetRecall()
-		m.updateThreadOverlays()
-		return
-	}
-	m.input = []rune(text)
-	m.inputPos = len(m.input)
-	m.inputHistory.ResetRecall()
-	m.updateInputOverlays()
-	m.maybeActivateChannelPickerFromInput()
-}
-
-func (m *channelModel) activeInputString() string {
-	if m.focus == focusThread && m.threadPanelOpen {
-		return string(m.threadInput)
-	}
-	return string(m.input)
-}
-
-func (m *channelModel) insertAcceptedMention(mention string) {
-	if m.focus == focusThread && m.threadPanelOpen {
-		m.threadInputHistory.ResetRecall()
-		m.threadInput, m.threadInputPos = channelui.ReplaceMentionInInput(m.threadInput, m.threadInputPos, mention)
-		m.updateThreadOverlays()
-		return
-	}
-	m.inputHistory.ResetRecall()
-	m.input, m.inputPos = channelui.ReplaceMentionInInput(m.input, m.inputPos, mention)
-	m.updateInputOverlays()
-}
-
-func (m *channelModel) restoreMainSnapshot(snapshot channelui.Snapshot) {
-	m.input = append([]rune(nil), snapshot.Input...)
-	m.inputPos = channelui.NormalizeCursorPos(m.input, snapshot.Pos)
-	m.updateInputOverlays()
-}
-
-func (m *channelModel) restoreThreadSnapshot(snapshot channelui.Snapshot) {
-	m.threadInput = append([]rune(nil), snapshot.Input...)
-	m.threadInputPos = channelui.NormalizeCursorPos(m.threadInput, snapshot.Pos)
-	m.updateThreadOverlays()
-}
-
-func composerMotionKey(msg tea.KeyMsg) (string, bool) {
-	if msg.Alt && len(msg.Runes) == 1 {
-		switch msg.Runes[0] {
-		case 'h':
-			return "alt+h", true
-		case 'l':
-			return "alt+l", true
-		case 'b':
-			return "alt+b", true
-		case 'w':
-			return "alt+w", true
-		case '0':
-			return "alt+0", true
-		case '$':
-			return "alt+$", true
-		}
-	}
-	switch msg.String() {
-	case "ctrl+a", "ctrl+e", "ctrl+b", "ctrl+f", "left", "right", "alt+h", "alt+l", "alt+b", "alt+w", "alt+0", "alt+$":
-		return msg.String(), true
-	default:
-		return "", false
-	}
-}
-
-func composerInsertRunes(msg tea.KeyMsg) []rune {
-	if msg.Type == tea.KeySpace || msg.String() == " " {
-		return []rune{' '}
-	}
-	if msg.Alt {
-		return nil
-	}
-	if len(msg.Runes) > 0 {
-		return msg.Runes
-	}
-	return nil
-}
+// Composer overlay management + composer motion/insert helpers moved
+// to channel_input.go.
 
 // runActiveCommand, runCommand, and maybeActivateChannelPickerFromInput moved
 // to channel_commands.go.
